Move ContextIO and ContextRequest to library
Move ContextIO (channel-based IO handle for context functions) and
ContextRequest from the flowrcli binary to the library module
flowr/src/lib/context_io.rs. The flowrcli binary re-exports from
the library.

This enables the context proxy implementation (which lives in the
library) to relay context function calls back to the origin
coordinator through the same ContextIO channels.

Add unit tests for ContextIO send_and_receive and send_no_reply.
